Added function to check if user has a basket, package, or checkoutUrl set.
Used to help validate and clear cookies on login if they don't match on our store page. This is for efficiency as this approach only clears the cookies fi they don't match the user's profile, reducing unnecessary operations. It also allows for user continuitity as the users can maintain their session details (basket and package) across logins if the data is still valid, providing a smoother experience.